Close the remaining medium findings
Uploads took their stored extension from the client's filename while only
the declared content type was validated, so a file announced as image/png
could be stored as .svg and served back as image/svg+xml from the API
origin. The extension now comes from the type that was actually checked.

Deactivating a tax rate did not stop it being used — anything still holding
the id kept creating documents at the old rate. Document list queries pulled
every line, lot allocation and serial for the page despite the list response
containing none of them. Stock movements gained the index the ledger poster
looks them up by on every posting and void. Sending an explicit null issue
date returned a 500 rather than a validation error, and stock-on-hand
declared float, rounding exact quantities on the way out.

+++ backend/app/modules/documents/router.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+from decimal import Decimal
 from typing import Annotated
 
 from fastapi import APIRouter, Depends, Query, Response, status
@@ -57,7 +58,7 @@ def sellable_items(
     return svc.sellable_items(membership.org_id, search, limit, warehouse_id)
 
 
-@router.get("/stock-on-hand", response_model=dict[int, float])
+@router.get("/stock-on-hand", response_model=dict[int, Decimal])
 def stock_on_hand(
     product_ids: list[int] = Query(default=[]),
     warehouse_id: int | None = None,
